pyproject.toml
[tool.poetry]
name = "zookeeper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "2.2.2"
selenium = "4.23.1"
sqlalchemy = "2.0.32"
fastapi = "0.112.2"
dbt-core = "1.8.5"
dbt-postgres = "1.8.2"
psycopg2 = "2.9.9"
confluent-kafka = "2.5.0"
requests = "2.32.3"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
